Add alex-shpak/hugo-book theme
This theme comes pretty close to what I'd like to see. It's not the
final product. It's simple, but it's still too complex for our needs.
We don't need to bundle JavaScript for search or mermaid with our site,
for example. Also, for longevity, we don't want to rely on any external
resources, if we can help it. So this theme will need to be pared down
and incorporated into this repo. But also, I do not want to get too
fixated on theming, when the basic content hasn't been migrated yet.

https://github.com/alex-shpak/hugo-book
https://themes.gohugo.io/hugo-book/